Split useSupabaseData into domain-specific hooks
To improve maintainability and separate concerns, the large
`useSupabaseData.ts` hook file has been split into three
domain-specific files:

- `useEvents.ts`: Core event fetching and mutations
- `useAttendances.ts`: Attendance records and joining logic
- `useProfile.ts`: User profile fetching

All components and pages importing these hooks have been updated
to reference the new paths. The test file has also been renamed
and updated to reflect the changes.
