+++ src/__tests__/hooks/supabase.test.tsx
@@ -7,9 +7,11 @@ import {
   fetchEvents,
   useCreateEvent,
   useDeleteEvent,
+} from "@/hooks/useEvents";
+import {
   useJoinEvent,
   useLeaveEvent,
-} from "@/hooks/useSupabaseData";
+} from "@/hooks/useAttendances";
 import { createQueryStub, supabaseStub } from "@/test-utils/supabase";
 
 const createWrapper = () => {
@@ -24,7 +26,7 @@ const createWrapper = () => {
   return { client, wrapper };
 };
 
-describe("useSupabaseData", () => {
+describe("supabase hooks", () => {
   it("fetches events with the organizer filter and descending order", async () => {
     const result = { data: [{ id: "event-1" }], error: null };
     const query = createQueryStub({ baseResult: result });
@@ -186,5 +188,8 @@ describe("useSupabaseData", () => {
     });
 
     invalidateSpy.mockRestore();
+  });
+});
+validateSpy.mockRestore();
   });
 });
