+++ src/hooks/useEvents.ts
@@ -9,21 +9,19 @@ import {
 import { supabase } from "@/integrations/supabase/client";
 import type { Database } from "@/integrations/supabase/types";
 
-type EventRow = Database["public"]["Tables"]["events"]["Row"];
-type ProfileRow = Database["public"]["Tables"]["profiles"]["Row"];
-type AttendanceRow = Database["public"]["Tables"]["attendances"]["Row"];
+export type EventRow = Database["public"]["Tables"]["events"]["Row"];
 
-type EventIdentifier = {
+export type EventIdentifier = {
   id?: string;
   slug?: string;
 };
 
-const normalizeEventIdentifier = (identifier?: EventIdentifier) => ({
+export const normalizeEventIdentifier = (identifier?: EventIdentifier) => ({
   id: identifier?.id ?? undefined,
   slug: identifier?.slug ?? undefined,
 });
 
-const ensureEventIdentifier = (identifier: EventIdentifier) => {
+export const ensureEventIdentifier = (identifier: EventIdentifier) => {
   if (!identifier.id && !identifier.slug) {
     throw new Error("An event id or slug is required");
   }
@@ -72,13 +70,13 @@ export const useEvent = (identifier?: EventIdentifier) => {
   });
 };
 
-type UseEventsOptions = {
+export type UseEventsOptions = {
   organizerId?: string;
   shortCode?: string;
   enabled?: boolean;
 };
 
-const normalizeEventsOptions = (options?: UseEventsOptions) => ({
+export const normalizeEventsOptions = (options?: UseEventsOptions) => ({
   organizerId: options?.organizerId ?? undefined,
   shortCode: options?.shortCode ?? undefined,
 });
@@ -132,143 +130,6 @@ export const useMyEvents = (
   });
 };
 
-type UseAttendancesOptions = {
-  eventId?: string;
-  userId?: string;
-  includeProfiles?: boolean;
-  includeEvents?: boolean;
-  enabled?: boolean;
-};
-
-type AttendanceRecord = AttendanceRow & {
-  profiles?: ProfileRow | null;
-  events?: EventRow | null;
-};
-
-const normalizeAttendancesOptions = (options: UseAttendancesOptions) => ({
-  eventId: options.eventId ?? undefined,
-  userId: options.userId ?? undefined,
-  includeProfiles: options.includeProfiles ?? false,
-  includeEvents: options.includeEvents ?? false,
-});
-
-export const attendancesQueryKey = (options: UseAttendancesOptions) => [
-  "attendances",
-  normalizeAttendancesOptions(options),
-];
-
-export const fetchAttendances = async (
-  options: UseAttendancesOptions,
-): Promise<AttendanceRecord[]> => {
-  if (!options.eventId && !options.userId) {
-    throw new Error("eventId or userId must be provided");
-  }
-
-  const normalized = normalizeAttendancesOptions(options);
-
-  const selectParts = ["*"];
-
-  if (normalized.includeProfiles) {
-    selectParts.push("profiles(*)");
-  }
-
-  if (normalized.includeEvents) {
-    selectParts.push("events(*)");
-  }
-
-  const selectClause = selectParts.join(", ");
-
-  let query = supabase.from("attendances").select(selectClause);
-
-  if (normalized.eventId) {
-    query = query.eq("event_id", normalized.eventId);
-  }
-
-  if (normalized.userId) {
-    query = query.eq("user_id", normalized.userId);
-  }
-
-  const { data, error } = await query;
-
-  if (error) {
-    throw error;
-  }
-
-  return (data ?? []) as AttendanceRecord[];
-};
-
-export const useAttendances = (options: UseAttendancesOptions) => {
-  const enabled =
-    (options.enabled ?? true) && Boolean(options.eventId || options.userId);
-
-  return useQuery({
-    queryKey: attendancesQueryKey(options),
-    enabled,
-    queryFn: () => fetchAttendances(options),
-  });
-};
-
-export const useUpcoming = (
-  userId?: string,
-  options?: { enabled?: boolean },
-): UseQueryResult<EventRow[], Error> => {
-  const enabled = options?.enabled ?? Boolean(userId);
-
-  return useQuery({
-    queryKey: ["upcoming-events", userId],
-    enabled,
-    queryFn: async () => {
-      if (!userId) {
-        throw new Error("User id required to fetch upcoming events");
-      }
-
-      const records = await fetchAttendances({
-        userId,
-        includeEvents: true,
-      });
-
-      return (records ?? [])
-        .map((attendance) => attendance.events)
-        .filter(Boolean) as EventRow[];
-    },
-  });
-};
-
-export const profileQueryKey = (userId?: string) => ["profile", userId];
-
-const fetchProfile = async (userId: string): Promise<ProfileRow | null> => {
-  const { data, error } = await supabase
-    .from("profiles")
-    .select("*")
-    .eq("id", userId)
-    .maybeSingle();
-
-  if (error) {
-    throw error;
-  }
-
-  return data ?? null;
-};
-
-export const useMyProfile = (userId?: string) => {
-  return useQuery({
-    queryKey: profileQueryKey(userId),
-    enabled: Boolean(userId),
-    queryFn: () => fetchProfile(userId!),
-  });
-};
-
-type JoinEventVariables = {
-  eventId: string;
-  userId: string;
-  source?: string;
-};
-
-type LeaveEventVariables = {
-  eventId: string;
-  userId: string;
-};
-
 type CreateEventVariables = Database["public"]["Tables"]["events"]["Insert"];
 
 type UpdateEventVariables = {
@@ -286,48 +147,6 @@ type DeletedEventRecord = {
   id: string;
   slug: string;
   organizer_id: string;
-};
-
-export const useJoinEvent = () => {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: async ({ eventId, userId, source }: JoinEventVariables) => {
-      const { error } = await supabase.from("attendances").insert({
-        event_id: eventId,
-        user_id: userId,
-        source: source ?? "manual",
-      });
-
-      if (error) {
-        throw error;
-      }
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["attendances"] });
-    },
-  });
-};
-
-export const useLeaveEvent = () => {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: async ({ eventId, userId }: LeaveEventVariables) => {
-      const { error } = await supabase
-        .from("attendances")
-        .delete()
-        .eq("event_id", eventId)
-        .eq("user_id", userId);
-
-      if (error) {
-        throw error;
-      }
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["attendances"] });
-    },
-  });
 };
 
 export const useCreateEvent = () => {
@@ -464,5 +283,3 @@ export const fetchEventsWithClient = (
     queryKey: eventsQueryKey(options),
     queryFn: () => fetchEvents(options),
   });
-
-export type { EventRow, ProfileRow, AttendanceRecord };
