+++ src/pages/event/components/AttendeeList.tsx
@@ -2,7 +2,7 @@ import { Card, CardContent, CardHeader } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Linkedin, Users } from "lucide-react";
-import type { ProfileRow } from "@/hooks/useSupabaseData";
+import type { ProfileRow } from "@/hooks/useProfile";
 import { TEXT } from "@/constants/text";
 import { LINKEDIN } from "@/constants/brands";
 import { toast } from "sonner";
